Made suggeted changes
